+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Understand comments.
+
 - Add `nomis/ec-base-priority-for-overlays`.